Add files via upload
NumPy is a fundamental Python library for scientific computing. 
It provides powerful tools for handling large, multi-dimensional arrays and matrices, along with a collection of high-level mathematical functions to operate on these arrays efficiently. 
NumPy is widely used in data science, machine learning, statistics, and numerical analysis because of its speed and ease of integration with other libraries like Pandas, SciPy, and scikit-learn.